requirements.txt
altgraph==0.17.4
autopep8==2.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
django==5.2
djangorestframework==3.16.0
flake8==7.1.2
flake8-html==0.4.3
fonttools==4.54.1
gunicorn==23.0.0
iniconfig==2.0.0
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.2
mccabe==0.7.0
numpy==2.0.2
packaging==24.2
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.19.1
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pymongo==4.11.3
pyparsing==3.2.0
pytest==8.3.4
pytest-html==4.1.1
pytest-metadata==3.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytoolconfig==1.3.1
pywin32-ctypes==0.2.3
rope==1.13.0
setuptools==75.8.0
six==1.16.0
django==4.2.20
pymongo==4.11.3
